requirements.txt
firebase_admin==6.8.0
Flask==3.1.1
pandas==2.2.3
protobuf==6.31.0
python-dotenv==1.1.0
waitress==3.0.2
